requirements.txt


blinker==1.6.2
click==8.1.4
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1

Flask-JWT-Extended==4.4.2
Flask-Mail==0.9.1
Flask-Security==3.0.0

celery==5.2.3
redis==4.3.4
marshmallow==3.19.0
celery[redis]==5.2.3

Flask-RESTful
Flask-Migrate 


email-validator

flask_bootstrap

flask_caching

pdfkit